benchmark - this version works correctly, builds correctly and runs correctly on a deployed server
Note... I previously ran install for these:

    "can": "^3.3.1",
    "steal-stache": "^3.0.5"
    "steal": "^1.0.9",
    "steal-tools": "^1.0.2"

but in the wrong directory. Later, I ran them in a different directory and it broke things.
I reverted to this version of code and removed those dependencies from the erroneous
package.json but did not remove other files. Presumably the v3 files mentioned above are there someplace.