a few cosmetic tweaks over the last few months, also fixed IE problem with duplicate properties
turns out that DoneJS, when confronted with a view model that looks like this:

blah:{
	define:{
		xxx:'something'
	}
	xxx:'another thing'
}

somehow promotes them into the same object.

In strict mode, IE considers defining the same property to be bad. (I agree.)

I had this situation (the second one was a note from an earlier era). I fixed it.